Consolidate tmpdir implementations, include Windows
